+++ manager/transactionManager.go
@@ -2,78 +2,68 @@ package manager
 
 import(
 	"ct-budget-manager/model"
-	"net/http"
-	"encoding/json"
-	"io"
-	"io/ioutil"
 	"time"
 	"gopkg.in/mgo.v2/bson"
-	"github.com/gorilla/mux"
 )
 
 /*
-* POST request to create a new Transaction object
-*
-* @param string wording
-* @param string description
-* @param int amount
+* @param string budgetId
+* @param interface{} wording
+* @param interface{} description
+* @param interface{} tType
+* @param interface{} sector
+* @param interface{} amount
+* @parem *model.Transaction
 */
-func CreateTransaction(w http.ResponseWriter, r *http.Request) {
-	var t model.Transaction
-	body, err := ioutil.ReadAll(io.LimitReader(r.Body, 1048576))
-	CheckError(err)
-	err = r.Body.Close()
-	CheckError(err)
-	if err := json.Unmarshal(body, &t); err != nil {
-		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-		w.WriteHeader(http.StatusUnprocessableEntity)
-		if err := json.NewEncoder(w).Encode(err); err != nil {
-    			panic(err)
+func CreateTransaction(budgetId string, wording interface{}, description interface{}, tType interface{}, sector interface{}, amount interface{}) *model.Transaction {
+	sectorName := sector.(map[string]interface{})["name"]
+
+	transaction := model.Transaction{
+		Id: bson.NewObjectId(),
+		Wording: wording.(string),
+		Description: description.(string),
+		Type: tType.(bool),
+		Sector: model.Sector{Name: sectorName.(string)},
+		Amount: amount.(float64),
+		CreatedAt: time.Now(),
+	}
+	if !AddTransactionToBudget(budgetId, transaction) {
+		return nil
+	}
+	return &transaction
+}
+
+/*
+* @param string id
+* @return *model.Transaction
+*/
+func GetTransaction(id string) *model.Transaction {
+	var raw interface{}
+
+	if err := MongoDBConnection.
+	DB(MongoDBName).
+	C("budget").
+	Find(bson.M{"transactions._id": bson.ObjectIdHex(id)}).
+	Select(bson.M{"_id": 0, "transactions.$": 1}).
+	One(&raw); err != nil {
+		if err.Error() == "not found" {
+			return nil
 		}
+		panic(err)
 	}
-
-	err = json.Unmarshal(body, &t)
-
-	t.Id = bson.NewObjectId()
-	t.Date = time.Now()
-
-  	err = MongoDBConnection.DB(MongoDBName).C("budget").Insert(t)
-	CheckError(err)
-
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.WriteHeader(http.StatusCreated)
-	err = json.NewEncoder(w).Encode(t)
-	CheckError(err)
+	var data map[string]model.Transactions
+	bytes, _ := bson.Marshal(raw)
+	bson.Unmarshal(bytes, &data)
+	return &data["transactions"][0]
 }
 
-func GetTransaction(w http.ResponseWriter, r *http.Request) {
-
-	var t model.Transaction
-
-	vars := mux.Vars(r)
-
-	if !bson.IsObjectIdHex(vars["id"]) {
-		w.WriteHeader(http.StatusNotFound)
-		return
+/*
+* @return model.Transactions
+*/
+func GetTransactions() model.Transactions {
+	transactions := make(model.Transactions, 0)
+	if err := MongoDBConnection.DB(MongoDBName).C("budget").Find(nil).All(&transactions); err != nil {
+		panic(err)
 	}
-
-	err := MongoDBConnection.DB(MongoDBName).C("budget").FindId(bson.ObjectIdHex(vars["id"])).One(&t)
-	CheckError(err)
-
-	err = json.NewEncoder(w).Encode(t)
-	CheckError(err)
-
-}
-
-func GetTransactions(w http.ResponseWriter, r *http.Request) {
-
-	t := make(model.Transactions, 0)
-	err := MongoDBConnection.DB(MongoDBName).C("budget").Find(nil).All(&t)
-	CheckError(err)
-
-	w.Header().Set("Access-Control-Allow-Origin", "*")
-
-	err = json.NewEncoder(w).Encode(t)
-	CheckError(err)
-
+	return transactions
 }