+++ model/transaction.go
@@ -12,8 +12,8 @@ type(
 		Description string `json:"description"`
 		Type bool `json:"type"`
 		Sector Sector `json:"sector"`
-		Amount int `json:"amount"`
-		Date time.Time `json:"date"`
+		Amount float64 `json:"amount"`
+		CreatedAt time.Time `json:"created_at"`
 	}
 	Transactions []Transaction
 )