+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Budget details
 - Mongo database name configuration
 
+### Changed
+- Transactions are now bound to a budget
+- Transactions amount are now float type
+- Transaction "Date" field is now "CreatedAt"
+
 ## [0.0.1] - 2017-09-14
 ### Added
 - API Docker container configuration